fixed merge conflits
