+++ test/unit/contamination_test/test/load_contamination_a_16.cpp
@@ -29,6 +29,7 @@
 #include "detail/load_contamination.hpp"
 #include "kernel_generator.hpp"
 #include "unit_test.hpp"
+#include <rocwmma/internal/vector.hpp>
 
 namespace rocwmma
 {
